refactor: extract WarningsCallout to shared component
